requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
librosa
Pillow
python-multipart